+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -20,8 +20,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 424;
-constexpr std::size_t translation_table_string_table_size = 75572;
+constexpr std::uint16_t translation_table_mapping_table_size = 425;
+constexpr std::size_t translation_table_string_table_size = 75610;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -293,6 +293,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "missing body for 'if' statement"sv,
           "missing body for 'switch' statement"sv,
           "missing body for TypeScript interface"sv,
+          "missing body for TypeScript namespace"sv,
           "missing body for catch clause"sv,
           "missing body for class"sv,
           "missing body for do-while loop"sv,